skip report notif deletions as admins/staff/mods can still see the posts

+++ app/activitypub/util.py
@@ -1620,6 +1620,9 @@ def delete_post_or_comment(deletor, to_delete, store_ap_json, request_json, reas
             # remove any notifications about the post
             notifs = db.session.query(Notification).filter(Notification.targets.op("->>")("post_id").cast(Integer) == to_delete.id)
             for notif in notifs:
+                # dont delete report notifs
+                if notif.notif_type == NOTIF_REPORT or notif.notif_type == NOTIF_REPORT_ESCALATION:
+                    continue
                 db.session.delete(notif)
             db.session.commit()
         elif isinstance(to_delete, PostReply):
